+++ Classes/steam/packets/rcon/SCRCONPacket.m
@@ -7,10 +7,12 @@
 
 #import "SCRCONPacket.h"
 
-
 @implementation SCRCONPacket
 
--(id) initWithRequestId:(unsigned long) packetRequestId andHeader:(unsigned long) rconHeader andData:(NSData*) rconData {
+- (id)initWithRequestId:(uint32_t)packetRequestId
+              andHeader:(uint32_t)rconHeader
+                andData:(NSData *)rconData
+{
     self = [super initWithData:rconData];
     header = rconHeader;
     requestId = packetRequestId;
@@ -18,22 +20,23 @@
     return self;
 }
 
-
--(NSData*) getData {
+- (NSData *)getData
+{
     NSUInteger dataLength = [contentData getLength];
-    NSMutableData* bytes = [[NSMutableData alloc] initWithLength:0];
+    NSMutableData *bytes = [[NSMutableData alloc] initWithLength:0];
     unsigned int packetLength = dataLength + 10;
-    
-    [bytes appendData: [NSData dataWithBytes:(uint8_t*) &packetLength length:4]];
-    [bytes appendData: [NSData dataWithBytes:(uint8_t*) &requestId length:4]];
-    [bytes appendData: [NSData dataWithBytes:(uint8_t*) &header length:4]];
-    [bytes appendData: [contentData array]];
+
+    [bytes appendData:[NSData dataWithBytes:(uint8_t *)&packetLength length:4]];
+    [bytes appendData:[NSData dataWithBytes:(uint8_t *)&requestId length:4]];
+    [bytes appendData:[NSData dataWithBytes:(uint8_t *)&header length:4]];
+    [bytes appendData:[contentData array]];
     [bytes increaseLengthBy:2];
-    
+
     return bytes;
 }
 
--(unsigned long) getRequestId {
+- (unsigned long) getRequestId
+{
     return requestId;
 }
